moved to using 1 dir for a pdf instead of seperating it across 3 dirs

+++ util.py
@@ -55,19 +55,14 @@ def reset() -> None: #reset the db
         "CREATE TABLE registry (" #make a new one
             "id INTEGER GENERATED ALWAYS AS IDENTITY PRIMARY KEY," #id of the pdf
             "name VARCHAR(255) NOT NULL," #name for the pdf
-            "pdf_path TEXT," #path to the pdf file
-            "idx_path TEXT," #path to the index file
-            "res_path TEXT," #path to the results file
+            "dir TEXT," #path to the directory
             "timestamp TIMESTAMP DEFAULT CURRENT_TIMESTAMP" #timestamp of when the pdf was added
         ");")
     con.commit()
     cur.close()
     con.close()
-    if os.path.exists(STORAGE): #remove the directory if it exists
-        shutil.rmtree(STORAGE)
-    os.makedirs(os.path.join(STORAGE, "pdf"), exist_ok=True) #create the directory for storing pdfs
-    os.makedirs(os.path.join(STORAGE, "idx"), exist_ok=True) #create the directory for storing the AI's vector stores
-    os.makedirs(os.path.join(STORAGE, "res"), exist_ok=True) #create the directory for storing the AI's responses
+    if os.path.exists(STORAGE): shutil.rmtree(STORAGE) #remove the directory if it exists
+    os.makedirs(os.path.join(STORAGE, "notebooks"), exist_ok=True) #make directory for pdfs
 
 def get_pdfs() -> tuple: #get all pdfs from db
     con = get_db_connection()
@@ -85,15 +80,17 @@ def upload_pdfs(list: list) -> None: #upload a list of pdfs
         cur.execute("INSERT INTO registry (name) VALUES (%s) RETURNING id", (file.filename,))
         id = cur.fetchone()[0] #add the name to the registry to generate an id
         fname = os.path.splitext(secure_filename(f"{id}{file.filename}"))[0] #generate a unique name
-        pdf_path = os.path.join(f"{STORAGE}/pdf", f"{fname}.pdf") #generate the path to save the pdf
-        idx_path = os.path.join(f"{STORAGE}/idx", fname) #generate the path to save the index
-        res_path = os.path.join(f"{STORAGE}/res", f"{fname}.json") #generate the path to save the results
+        dir = os.path.join(STORAGE, "notebooks", fname) #create the name for the directory
+        os.makedirs(dir, exist_ok=True) #actually make the directory
+        pdf_path = os.path.join(dir, "source.pdf") #generate the path to save the pdf
+        idx_path = os.path.join(dir, "idx") #generate the path to save the index
+        res_path = os.path.join(dir, "res.json") #generate the path to save the results
         file.save(pdf_path) #save the pdf to the path
         os.makedirs(idx_path, exist_ok=True) #make the directory for the index
         idx = VectorStoreIndex.from_documents(SimpleDirectoryReader(input_files=[pdf_path]).load_data()) #create the index from the pdf
         idx.storage_context.persist(persist_dir=idx_path) #save the index to the path
         with open(res_path, 'w') as f: json.dump({}, f) #initialize the json as empty
-        cur.execute("UPDATE registry SET pdf_path = %s, idx_path = %s, res_path = %s WHERE id = %s", (pdf_path, idx_path, res_path, id)) #add the name and paths to the registry
+        cur.execute("UPDATE registry SET dir = %s WHERE id = %s", (dir, id)) #add the name and paths to the registry
     con.commit()
     cur.close()
     con.close()
@@ -101,14 +98,9 @@ def upload_pdfs(list: list) -> None: #upload a list of pdfs
 def delete_pdf(id: int) -> None: #delete a pdf
     con = get_db_connection()
     cur = con.cursor()
-    cur.execute("DELETE FROM registry WHERE id = %s RETURNING (pdf_path, idx_path, res_path)", (id,)) #remove the pdf from the registry
-    pathes = cur.fetchone() #get the paths to the pdf, index, and results file
-    if os.path.exists(pathes[0]): #remove the pdf file
-        os.remove(pathes[0])
-    if os.path.exists(pathes[1]): #remove the index directory
-        shutil.rmtree(pathes[1])
-    if os.path.exists(pathes[2]): #remove the results file
-        os.remove(pathes[2])
+    cur.execute("DELETE FROM registry WHERE id = %s RETURNING dir", (id,)) #remove the pdf from the registry
+    dir = cur.fetchone()[0] #get the paths to the pdf, index, and results file
+    if os.path.exists(dir): shutil.rmtree(dir) #remove the directory
     con.commit()
     cur.close()
     con.close()
@@ -116,8 +108,8 @@ def delete_pdf(id: int) -> None: #delete a pdf
 def get_idx(id: int) -> VectorStoreIndex: #get an idx
     con = get_db_connection()
     cur = con.cursor()
-    cur.execute("SELECT idx_path FROM registry WHERE id = %s", (id,))
-    idx_path = cur.fetchone()[0] #get the path of the index to load
+    cur.execute("SELECT dir FROM registry WHERE id = %s", (id,))
+    idx_path = os.path.join(cur.fetchone()[0], "idx") #get the path of the index to load
     cur.close()
     con.close()
     return load_index_from_storage(StorageContext.from_defaults(persist_dir=idx_path)) #load the index from that path and return it
@@ -125,8 +117,8 @@ def get_idx(id: int) -> VectorStoreIndex: #get an idx
 def set_res(id: int, task: Task, res: str) -> None: #set a result
     con = get_db_connection()
     cur = con.cursor()
-    cur.execute("SELECT res_path FROM registry WHERE id = %s", (id,))
-    res_path = cur.fetchone()[0] #get the path of the results file to save to
+    cur.execute("SELECT dir FROM registry WHERE id = %s", (id,))
+    res_path = os.path.join(cur.fetchone()[0], "res.json") #get the path of the results file to save to
     cur.close()
     con.close()
     with open(res_path, 'r') as f: data = json.load(f) #get the current state
@@ -136,27 +128,29 @@ def set_res(id: int, task: Task, res: str) -> None: #set a result
 def get_res(id: int, task: Task = None) -> str: #get a result
     con = get_db_connection()
     cur = con.cursor()
-    cur.execute("SELECT res_path FROM registry WHERE id = %s", (id,))
-    res_path = cur.fetchone()[0] #get the path of the results file to read from
+    cur.execute("SELECT dir FROM registry WHERE id = %s", (id,))
+    res_path = os.path.join(cur.fetchone()[0], "res.json") #get the path of the results file to read from
     cur.close()
     con.close()
     with open(res_path, 'r', encoding='utf-8') as f: data = json.load(f) #get the current state
     return data.get(task.name.lower()) if task else data #return value for the key of the task, if no task specified return the whole thing
 
-def query(id: int, task: Task) -> str: #query the llm to do a task
+def query(idx: VectorStoreIndex, task: Task) -> str: #query the llm to do a task
     return RetrieverQueryEngine.from_args(
-        get_idx(id).as_retriever(similarity_top_k=30), #get chunks from the pdf
+        idx.as_retriever(similarity_top_k=30), #get chunks from the pdf
         response_mode="compact",
         text_qa_template=PromptTemplate(PROMPTS['base'].format( #use the prompt that includes the rubric, context from pdf, and the task
             rubric_ref = PROMPTS['rubric_ref'],
             context_str = "{context_str}",
             query_str="{query_str}"
-        ))
-    ).query(task.value).response #enter the task as the query
+    ))).query(task.value).response #enter the task as the query
 
 def query_and_write_all(id : int): #does all the tasks for a pdf
+    print(f'loading idx')
+    idx = get_idx(id)
     for t in Task:
-        set_res(id, t, query(id, t))
+        print(f'starting {t.name}')
+        set_res(id, t, query(idx, t))
         print(f'finished {t.name}')
 
 if __name__ == "__main__":
